+++ src/main/java/forestry/api/modules/ForestryModuleIds.java
@@ -22,4 +22,5 @@ public class ForestryModuleIds {
 	public static final ResourceLocation SORTING = ForestryConstants.forestry("sorting");
 	public static final ResourceLocation CULTIVATION = ForestryConstants.forestry("cultivation");
 	public static final ResourceLocation WORKTABLE = ForestryConstants.forestry("worktable");
+	public static final ResourceLocation CURIOS = ForestryConstants.forestry("curios");
 }